# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Service Portal API",
    "version": "1.0.0",
...
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register an account",
        "description": "Create an account (AUTH-1 through AUTH-5, AUTH-16).\n\nDoes **not** sign the new user in: no cookie is set, and the caller is\nexpected to send them to the login form (design.md §2). Registration and\nauthentication stay separate events, which keeps this endpoint from being\na way to mint a session.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisteredUser"
                }
              }
            }
          },
          "409": {
            "description": "Returned when that email address is already registered (AUTH-2).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Returned when one or more supplied fields are invalid (XC-4).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Returned when the `X-Requested-With` header is missing, which middleware refuses before routing (XC-9).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Returned when the request failed for a reason the API does not attribute to the caller — `message` never carries internal detail and the traceback is logged server-side (XC-14).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Requested-With",
              "default": "XMLHttpRequest"
            },
            "description": "CSRF guard (XC-9). Any non-empty value. Enforced by middleware before routing, so a request without it is answered `403` without its body or its cookies being read — which is why it is absent from this operation's handler signature."
          }
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign in and start a session",
        "description": "Start a session (AUTH-6 through AUTH-8).",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionUser"
                }
              }
            }
          },
          "401": {
            "description": "Returned when the credentials are rejected — an unknown address, a wrong password and a deactivated account are answered identically (AUTH-7, AUTH-8).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Returned when one or more supplied fields are invalid (XC-4).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Returned when the `X-Requested-With` header is missing, which middleware refuses before routing (XC-9).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Returned when the request failed for a reason the API does not attribute to the caller — `message` never carries internal detail and the traceback is logged server-side (XC-14).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Requested-With",
              "default": "XMLHttpRequest"
            },
            "description": "CSRF guard (XC-9). Any non-empty value. Enforced by middleware before routing, so a request without it is answered `403` without its body or its cookies being read — which is why it is absent from this operation's handler signature."
          }
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "The signed-in user",
        "description": "Who is signed in (AUTH-14, AUTH-15).\n\nThe frontend's session bootstrap: cookie-based sessions live on the server,\nso this is how a freshly loaded page discovers whether it has one.\n\n``get_current_user`` reads the access-token cookie and nothing else, which\nis AUTH-15's requirement rather than an omission — a `/me` that quietly\nrefreshed would make an expired session indistinguishable from a live one,\nand refreshing is `/auth/refresh`'s job.",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionUser"
                }
              }
            }
          },
          "401": {
            "description": "Returned when the `access_token` cookie is absent, expired, or no longer names an active user — no distinction is drawn between those cases (XC-5).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Returned when the request failed for a reason the API does not attribute to the caller — `message` never carries internal detail and the traceback is logged server-side (XC-14).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessTokenCookie": []
          }
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Rotate the session cookies",
        "description": "Rotate the refresh token and reissue both cookies (AUTH-9 through AUTH-11).\n\nReads the refresh cookie directly rather than depending on\n``get_current_user``: the access token is expected to be expired by the\ntime anything calls this, which is the whole point of the endpoint.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionUser"
                }
              }
            }
          },
          "401": {
            "description": "Returned when the refresh cookie is absent, unknown, expired, already rotated, or belongs to a user who is gone or deactivated (AUTH-9 to AUTH-11).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Returned when the `X-Requested-With` header is missing, which middleware refuses before routing (XC-9).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Returned when the request failed for a reason the API does not attribute to the caller — `message` never carries internal detail and the traceback is logged server-side (XC-14).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "refreshTokenCookie": []
          }
        ],
        "parameters": [
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Requested-With",
              "default": "XMLHttpRequest"
            },
            "description": "CSRF guard (XC-9). Any non-empty value. Enforced by middleware before routing, so a request without it is answered `403` without its body or its cookies being read — which is why it is absent from this operation's handler signature."
          }
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "End the session",
        "description": "End the session (AUTH-12, AUTH-13). Always ``204``.\n\nDeliberately not gated on ``get_current_user``. AUTH-13 makes logout\nidempotent, and a ``401`` for a caller whose session had already expired\nwould fail their \"sign out\" click at the one moment it is least useful to\nargue about — and would leave the stale cookies sitting in their browser.\nClearing the cookies is unconditional for the same reason.\n\nReturns the ``Response`` object directly so the body is genuinely empty:\na ``204`` carrying a serialized `null` would be malformed.",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "description": "Returned when the `X-Requested-With` header is missing, which middleware refuses before routing (XC-9).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Returned when the request failed for a reason the API does not attribute to the caller — `message` never carries internal detail and the traceback is logged server-side (XC-14).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Requested-With",
              "default": "XMLHttpRequest"
            },
            "description": "CSRF guard (XC-9). Any non-empty value. Enforced by middleware before routing, so a request without it is answered `403` without its body or its cookies being read — which is why it is absent from this operation's handler signature."
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 255,
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "description": "``POST /auth/login`` body (AUTH-6, AUTH-7)."
      },
...
      "RegisterRequest": {
        "properties": {
          "first_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Last Name"
          },
          "email": {
            "type": "string",
            "maxLength": 255,
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "first_name",
          "last_name",
          "email",
          "password"
        ],
        "title": "RegisterRequest",
        "description": "``POST /auth/register`` body (AUTH-1, AUTH-3, AUTH-4, AUTH-16)."
      },
      "RegisteredUser": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "first_name",
          "last_name",
          "email",
          "role",
          "created_at"
        ],
        "title": "RegisteredUser",
        "description": "``POST /auth/register``'s ``201`` body (AUTH-1).\n\nField-for-field from design.md §2, in that order. Not built by extending\n``SessionUser``: the two overlap today by coincidence of what registration\nhappens to echo back, and tying them together would propagate any later\nchange to the session shape into the registration contract."
      },
...
      "SessionUser": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "first_name",
          "last_name",
          "role"
        ],
        "title": "SessionUser",
        "description": "Who the caller is — login, refresh and ``/auth/me`` all return this.\n\nOne shape for all three (design.md §2) so the frontend has a single\n\"current user\" type regardless of which call produced it. No `email`, no\ntimestamps, and — the point of the model existing — no `password_hash`."
      },
...
      "ErrorDetail": {
        "title": "ErrorDetail",
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "title": "Code",
            "type": "string",
            "enum": [
              "BAD_REQUEST",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "VALIDATION_ERROR",
              "INTERNAL_ERROR"
            ],
            "description": "Machine-readable error class. Closed set."
          },
          "message": {
            "title": "Message",
            "type": "string",
            "description": "Human-readable summary, safe to show a user. Never carries server internals."
          },
          "fields": {
            "title": "Fields",
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Per-field messages, keyed by field name. Present only on `VALIDATION_ERROR`."
          }
        }
      },
      "ErrorEnvelope": {
        "title": "ErrorEnvelope",
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "description": "The one shape every error response in this API uses."
      }
    },
    "securitySchemes": {
      "accessTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "Session access token. Set by `POST /api/v1/auth/login` and `POST /api/v1/auth/refresh` as an `httpOnly`, `Secure`, `SameSite=Lax` cookie with a 1 hour lifetime. Sent automatically by the browser; not readable from JavaScript and not settable by a client."
      },
      "refreshTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token",
        "description": "Opaque refresh token, 30 day lifetime, same cookie attributes as `access_token`. Read by `POST /api/v1/auth/refresh` only, which rotates it; replaying an already-rotated token revokes the whole family (AUTH-11)."
      }
    }
  }
}